Prefabs
Created Prefabs for all of the objects in level 1 so they can be used later. Also added more collectables and altered some code to fix the bonus scoring values added at the end of the level(s). Added a serialized field for the bonus time so it can be adjusted on each level.